# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunUsage": {
        "additionalProperties": false,
        "properties": {
          "by_agent": {
            "items": {
              "$ref": "#/components/schemas/UsageAgent"
            },
            "title": "By Agent",
            "type": "array"
          },
          "by_phase_model": {
            "additionalProperties": {
              "$ref": "#/components/schemas/UsageTotals"
            },
            "title": "By Phase Model",
            "type": "object"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "title": "Models",
            "type": "array"
          },
          "recorded": {
            "title": "Recorded",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "spend": {
            "title": "Spend",
            "type": "null"
          },
          "spend_basis": {
            "title": "Spend Basis",
            "type": "string"
          },
          "total": {
            "$ref": "#/components/schemas/UsageTotals"
          },
          "turns": {
            "default": 0,
            "title": "Turns",
            "type": "integer"
          },
          "workspace_id": {
            "default": "local",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "recorded",
          "total",
          "spend_basis"
        ],
        "title": "RunUsage",
        "type": "object"
      },
...
      "UsageAgent": {
        "additionalProperties": false,
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "jobs": {
            "default": 0,
            "title": "Jobs",
            "type": "integer"
          },
          "turns": {
            "default": 0,
            "title": "Turns",
            "type": "integer"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageTotals"
          }
        },
        "required": [
          "agent",
          "usage"
        ],
        "title": "UsageAgent",
        "type": "object"
      },
      "UsageTotals": {
        "additionalProperties": false,
        "properties": {
          "cache_read_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Read Tokens"
          },
          "cache_write_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Write Tokens"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          }
        },
        "title": "UsageTotals",
        "type": "object"
      },
      "UsageWindow": {
        "additionalProperties": false,
        "properties": {
          "models": {
            "items": {
              "type": "string"
            },
            "title": "Models",
            "type": "array"
          },
          "observed_at": {
            "title": "Observed At",
            "type": "string"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/UsageWindowRun"
            },
            "title": "Runs",
            "type": "array"
          },
          "runs_considered": {
            "default": 0,
            "title": "Runs Considered",
            "type": "integer"
          },
          "runs_recorded": {
            "default": 0,
            "title": "Runs Recorded",
            "type": "integer"
          },
          "since": {
            "title": "Since",
            "type": "string"
          },
          "spend": {
            "title": "Spend",
            "type": "null"
          },
          "spend_basis": {
            "title": "Spend Basis",
            "type": "string"
          },
          "total": {
            "$ref": "#/components/schemas/UsageTotals"
          },
          "turns": {
            "default": 0,
            "title": "Turns",
            "type": "integer"
          },
          "until": {
            "title": "Until",
            "type": "string"
          },
          "workspace_id": {
            "default": "local",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "since",
          "until",
          "observed_at",
          "total",
          "spend_basis"
        ],
        "title": "UsageWindow",
        "type": "object"
      },
      "UsageWindowRun": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "recorded": {
            "title": "Recorded",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "total": {
            "$ref": "#/components/schemas/UsageTotals"
          },
          "turns": {
            "default": 0,
            "title": "Turns",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "kind",
          "state",
          "recorded",
          "total"
        ],
        "title": "UsageWindowRun",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "sbxloop",
    "version": "snapshot"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/runs/{run_id}/usage": {
      "get": {
        "description": "Tokens and turns the run's backend reported, by persona and by\nphase and model; unknown values stay unknown.",
        "operationId": "get_run_usage_v1_runs__run_id__usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunUsage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Usage",
        "tags": [
          "usage"
        ]
      }
    },
...
    "/v1/usage": {
      "get": {
        "description": "Reported usage across the runs touched in a window (RFC 3339 or\nepoch bounds; the daemon's current calendar day when omitted; at most\n90 days wide).",
        "operationId": "get_usage_v1_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageWindow"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Usage",
        "tags": [
          "usage"
        ]
      }
    },
...
  }
}